Cleanup in Ball-classes. Making one abstract Ball-class and two sub-classes for each game.
